refactor:本一覧ページを分割してコンポーネント化した
