+++ components/BookCard/BookCard.module.scss
@@ -1,0 +1,20 @@
+.bookCard {
+  margin: auto;
+  margin-bottom: 10px;
+}
+
+.thumbnail {
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.returnButton {
+  background-color: #6b7280;
+  color: white;
+  font-weight: bold;
+  padding: 8px 16px;
+  border-radius: 4px;
+
+  &:hover {
+    background-color: #1d4ed8;
+  }
+}
