+++ components/LoginButton/LoginButton.tsx
@@ -1,0 +1,5 @@
+import { signIn } from "next-auth/react";
+
+export const LoginButton = () => {
+  return <button onClick={() => signIn("slack")}>Login</button>;
+};
