+++ package.json
@@ -19,7 +19,8 @@
     "postcss-simple-vars": "^7.0.1",
     "react": "^18",
     "react-dom": "^18",
-    "react-toastify": "^10.0.5"
+    "react-toastify": "^10.0.5",
+    "sass": "^1.76.0"
   },
   "devDependencies": {
     "@types/node": "^20.11.25",
