+++ types/book.ts
@@ -1,0 +1,9 @@
+export type Book = {
+  id: string;
+  title: string;
+  author: string;
+  description: string;
+  imageLink: string;
+  isLending: boolean;
+  publishedDate: string;
+};
